migrated classes in parameters.py to models.py
